добавлен функционал телеграм v1

+++ Gymopedia.Data/Repository/ClientToSessionRepository.cs
@@ -16,12 +16,12 @@ namespace Gymopedia.Data.Repository
             Context.ClientToSession.Add(ClientToSession);
         }
 
-        public async Task<ClientToSession?> Get(int clientId,int sessionId, CancellationToken cancellationToken)
+        public async Task<ClientToSession?> Get(long clientId, long sessionId, CancellationToken cancellationToken)
         {
             return await Context.ClientToSession.FirstOrDefaultAsync(o => o.Id == clientId, cancellationToken);
         }
 
-        public async Task<ClientToSession?> Delete(int clientId, int sessionId, CancellationToken cancellationToken)
+        public async Task<ClientToSession?> Delete(long clientId, long sessionId, CancellationToken cancellationToken)
         {
             ClientToSession? ClientToSession = await Context.ClientToSession.SingleOrDefaultAsync(o => o.Id == clientId, cancellationToken);
             if (ClientToSession == null)
