+++ Gymopedia.Domain/Models/ClientToSession.cs
@@ -4,7 +4,7 @@ namespace Gymopedia.Domain.Models
     public class ClientToSession
     {
         public int Id { get; set; }
-        public int SessionId { get; set; }
-        public int ClientId  { get; set; }
+        public long SessionId { get; set; }
+        public long ClientId  { get; set; }
     }
 }
